+++ tap_asana/streams/__init__.py
@@ -1,5 +1,9 @@
+import tap_asana.streams.portfolios
 import tap_asana.streams.projects
+import tap_asana.streams.sections
+import tap_asana.streams.stories
 import tap_asana.streams.tags
 import tap_asana.streams.tasks
+import tap_asana.streams.teams
 import tap_asana.streams.users
 import tap_asana.streams.workspaces